stock_report/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-report"
version = "0.1.0"
description = "Stock and ETF research report generator"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "PyYAML>=6.0",
    "requests>=2.31",
    "beautifulsoup4>=4.11",
]

[project.scripts]
stock-report = "stock_report.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
